+++ routes/authRoute.js
@@ -1,4 +1,4 @@
-const { renderSignUpPage, renderLoginPage, renderHomePage, signUpHandler, loginHandler, accessChekingHandler, logoutHandler } = require("../controllers/authRouterController");
+const { renderSignUpPage, renderLoginPage, renderHomePage, signUpHandler, loginHandler, accessChekingHandler, logoutHandler, refreshTokenHanlder } = require("../controllers/authRouterController");
 
 const router = require("express").Router();
 
@@ -6,7 +6,8 @@ router.route("/").get(renderHomePage);
 router.route("/login").get(renderLoginPage).post(loginHandler);
 router.route("/signup").get(renderSignUpPage).post(signUpHandler);
 router.route("/authcheck").get(accessChekingHandler);
-router.route("/logout").get(logoutHandler);
+router.route("/logout").post(logoutHandler);
+router.route("/refreshtoken").post(refreshTokenHanlder)