+++ utils/oAuthFunctions.js
@@ -1,9 +1,12 @@
 //? Initialize express server
+const { log } = require("console");
 const express = require("express");
 const app = express();
 
 //! OAUTH
 const OAuth2Server = require("oauth2-server");
+const { ErrorHandler } = require("./errorHandler");
+const { coloredLog } = require("./coloredLog");
 const Request = OAuth2Server.Request;
 const Response = OAuth2Server.Response;
 
@@ -14,18 +17,31 @@ app.oauth = new OAuth2Server({
 });
 
 let obtainToken = async (req, res) => {
-  var request = new Request(req);
-  var response = new Response(res);
+  try{
 
-  let token = await app.oauth.token(request, response).then((token) => {
-    // console.log("TOKEN ------- ", token);
+    var request = new Request(req);
+    var response = new Response(res);
+    
+    let token = await app.oauth.token(request, response).then((token) => {
+      // console.log("TOKEN ------- ", token);
+      return token;
+    });
+    
     return token;
-  });
-  return token;
+  } catch (error){
+    log("inside error", error)
+    throw new ErrorHandler(error.message, 401)
+  }
 };
 
 let authenticateRequest = async (req, res) => {
   try {
+
+    const {accessToken} = req.body
+
+    req.headers = {
+      authorization: `Bearer ${accessToken}`
+    }
     var request = new Request(req);
     var response = new Response(res);